adding grid component and size utilities
This adds two dependencies to the project:

- suitcss/components-grid
- suitcss/utils-size

These are included as the first additions to the `components` and `utils` style directories per the team CSS guidelines.

The included pattern documentation is a straightforward example that showcases responsive layout shifts, otherwise relying on the component docs themselves for more information.

This example would benefit from visual differentation of the columns, but rather than create some bespoke "demo" styles for this, I wanted to see if we either created one already (a "well" or "panel" pattern) or if other such  needs arise that we can address holistically.