initial color palette
Adding the color palette values from @nicolemors' branch, but also to the theme file. These are not applied anywhere yet, I'll save that for when we're actually plugging in the various patterns.

The variable names are loosely based on those in the SUITCSS example: https://github.com/suitcss/theme/blob/master/lib/theme.css