thumbnail component
The existing content and required features for the site will include a lot of user avatars and similar thumbnail imagery. This pattern applies an inner shadow as per the element collages, as well as subtle effects for states like `hover`, `focus` and `active`.

I've kept things pretty unopinionated to begin with, applying no `border-radius` by default. The `rounded` and `circle` modifiers can be used to customize the appearance. The collage only used circular avatars, but that seemed too dramatic for the default behavior... I figure we can always revisit and consolidate if we get tired of typing the modifiers.