Add WIP SVG icon component, task, examples
This is the first step to getting a robust and performant icon system in place.

I've chosen to use SVG for icons. They're a more semantic choice than icon fonts, offer more design flexibility, render more predictably, and easier to implement fallbacks for (IMO).

Contrary to the last couple of times I've implemented an SVG icon system, I decided not to set `fill: currentColor` this time. On past projects, I've encountered exceptions to this that have required some CSS overrides to accomodate. I think the icon designer should know which parts of the icon should or shouldn't inherit `currentColor`, so I've moved that into the icon assets themselves.

The icons are assembled into an SVG sprite using `gulp-svg-sprite`, my personal favorite tool for this sort of thing.

Right now there are only two icons, and those should probably be advised (they are currently 26px wide each, which is an odd measurement). We should also explore having a standalone "Icons" page that can be a larger reference of everything that's available.

To keep things simple for now, I'm referencing the SVG path directly, but that will fail in IE. We should follow up later to decide if we want to include a polyfill, or the SVG sprite in the `<body>`.