Add DRY arrow variations
Based on a technique detailed in this blog post: http://blog.cloudfour.com/our-svg-icon-process/

The reason I prefer this technique versus the transform one is that if any icon variations have semantic meaning in a given context, this makes it easier to identify those accordingly.

I think it also maps to the way we tend to think about icons without repeating stuff unnecessarily.