Update nav prototype with less flicker
The "animate page via transform" option in this prototype previously had a substantial "flicker" of visible content on-device. This version refactors it to avoid the flicker by calculating the element height in advance of its visibility.

The "animate page scroll" option has been removed as it had no perceivable benefits over the transform option.