Simplify base blockquote, tweak for CmsContent
It was difficult within the context of blog posts to differentiate blockquotes from surrounding content, particularly when they contained things like lists that had a similar left-padding treatment. Also, while our fancy `<cite>` styles looked nice in the style guide, they broke if anything other than paragraphs preceded them, which seems pretty fragile.

The challenge was that we're using blockquotes in a more neutral way on some marketing pages (for testimonial pull-quotes, for example), and I didn't want to make overly opinionated changes we'd have to create a pile of new utility classes for.

So I went with simplifying the root blockquote and adding a more specific treatment to the `CmsContent` class so it would be styled in that context.

See: #362