Add Storybook Build to CI Actions
This commit adds a new job to the GitHub Actions CI to run a Storybook
build. Previously, this wasn't happening, which led to CI passing even
though Storybook build failed on Netlify.

This also removes the preprocess step from the build job because the
build script already runs the preprocess step.