Add Short Date Format and Unlinking to Author
This PR adds two new features to the Author component. First, it adds
the `unlink` property, which will prevent a link from being rendered
even if one is provided as part of the Author object. This is useful
for situations where removing the link property from the Author is
a hassle. Second, it adds a `date_format` property, which can be set
to `short` to use abbreviate date formatting.