Use `overflow-wrap:anywhere` for Word Breaking
This PR changes our typography stylesheet from using the `break-word`
value of `overflow-wrap` to `anywhere`, which allows the line breaks
added to be considered when calculating min-content intrinsic sizes,
such as in CSS Grid layouts. Previously, long words were failing to
break in CSS Grid contexts like Card layouts.

It also updates many of our Prototypes to include a long string,
so we can test that it's working properly in more contexts.

Fixes #1729