Use `initialArgs` as Fallback for `args`
Some of our source code previews stopped displaying args passed to the
template recently. That seems to be due to the fact that the
`transformSource` function is called repeatedly, and `args` is not
always populated. It looks like `initialArgs` is a safe fallback, and
we already had several spots in the code doing so. This PR updates
a few stories to include the fallback that were missing it.

Fixes #1883