Add `is_sectioning` check to Card and Overview
This PR updates the template for the Card component and Overview object
to check if `tag_name` is a sectioning element like `article` or
`section`, and then change the default for `header_tag_element` and
`footer_tag_element` to either `header`/`footer` or `div`.

This avoids an issue when `header` elements are rendered in a `div`,
it causes useless "banner" landmarks to display in the VO Rotor.