Add `stylelint-scss@4.0.1` as a devDependency (#1993)
This PR adds `stylelint-scss@4.0.1` as a devDependency. Previously,
this was being implicitly added via the Cloud Four Stylelint config,
which specifies `stylelint-scss@^4.0.0` via Stylelint's recommended
SCSS config. However, there's a bug in v4.0.0 that was fixed in v4.0.1,
and we wanted that fix. Since v4.0.1 was released over 8 months ago
and the configs don't seem to be in a rush to bump the version, we're
doing it by hand. Someday when the recommended config bumps their
version, we could remove this devDependency.

Fixes #1890