Add CSS for Outlined style for WP Image block
This PR adds the styles for the new "Outlined" block style we added to
the Image block in the WordPress theme. Ideally, we would just add the
`u-border-small` class directly, but WordPress adds the class to the
outermost element — in this case, the `figure` — and we only want the
border to apply to the nested `img` element.

Fixes #862