Ground Nav v2 (#2194)
This PR updates the Ground Nav component to version two. This version
intentionally drops the "PWA Stats" use case, and so is a breaking change.

The intent of this redesign is largely to allow for two "feature" blocks
which could be configured. The layout is meant to handle zero, one, or two
features automatically.