Bugfix: Ground Nav Feature Widths
This PR fixes a bug where the width of the features for the Ground Nav
wasn't quite right. Now, when there are two features, they'll match
the width of the Ground Nav content, and when there's one feature,
it'll match the width of a `fill-pad` container in the prose-width
content above the Ground Nav.